refactor: don't have to build rocksdb when use all jsonrpc types
