+++ rpc/src/module/trace.rs
@@ -5,14 +5,13 @@ use ckb_protocol::RelayMessage;
 use ckb_shared::index::ChainIndex;
 use ckb_shared::shared::Shared;
 use ckb_shared::tx_pool::types::PoolEntry;
-use ckb_shared::tx_pool::TxTrace;
 use ckb_sync::NetworkProtocol;
 use ckb_traits::chain_provider::ChainProvider;
 use ckb_verification::TransactionError;
 use flatbuffers::FlatBufferBuilder;
 use jsonrpc_core::{Error, Result};
 use jsonrpc_derive::rpc;
-use jsonrpc_types::Transaction;
+use jsonrpc_types::{Transaction, TxTrace};
 use log::{debug, warn};
 use numext_fixed_hash::H256;
 use std::convert::TryInto;
