+++ shared/Cargo.toml
@@ -11,6 +11,7 @@ bincode = "1.1"
 ckb-core = { path = "../core" }
 ckb-chain-spec = { path = "../spec" }
 ckb-util = { path = "../util" }
+jsonrpc-types = { path = "../util/jsonrpc-types" }
 ckb-db = { path = "../db" }
 numext-fixed-hash = { version = "0.1", features = ["support_rand", "support_heapsize", "support_serde"] }
 numext-fixed-uint = { version = "0.1", features = ["support_rand", "support_heapsize", "support_serde"] }
