+++ shared/src/tx_pool.rs
@@ -3,7 +3,6 @@ pub mod trace;
 pub mod types;
 
 pub use self::pool::TxPool;
-pub use self::trace::TxTrace;
 pub use self::types::{
     OrphanPool, PendingQueue, PoolEntry, PoolError, StagingPool, StagingTxResult, TxPoolConfig,
 };
