+++ util/jsonrpc-types/src/lib.rs
@@ -4,6 +4,7 @@ mod bytes;
 mod cell;
 mod net;
 mod proposal_short_id;
+mod trace;
 
 pub use self::block_template::{
     BlockTemplate, CellbaseTemplate, TransactionTemplate, UncleTemplate,
@@ -12,4 +13,5 @@ pub use self::blockchain::{Block, Header, OutPoint, Transaction, UncleBlock};
 pub use self::bytes::Bytes;
 pub use self::cell::{CellOutputWithOutPoint, CellWithStatus};
 pub use self::net::{Node, NodeAddress};
+pub use self::trace::{Action, TxTrace};
 pub use jsonrpc_core::types::{error, id, params, request, response, version};
